Refs #7 - Added support for passworded channels
Some channels have a password requirement that must be sent in during
the join procedure.

Added a 'channel_passwords' config array index that, if present,
supplies the passwords for the channels defined in 'channels'.

The array is a parallel construction with the indices matching the
indices in 'channels'.  If it is omitted, backwards compatability is
preserved.